01_llm_clients/requirements.txt
python-dotenv
groq
openai